debug renaming of entities folder issue
